progress made in stub. refactored code & CLI
