+++ src/actions/trash_list.rs
@@ -1,78 +1,12 @@
-use std::{fs::{read_dir, read_to_string}, io::{Error, Result}, path::{Path, PathBuf}};
-use chrono::NaiveDateTime;
-use configparser::ini::Ini;
-use tabled::{settings::Style, Table, Tabled};
+use std::{
+    fs::read_dir, 
+    io::Result, 
+    path::Path
+};
+
+use tabled::{settings::Style, Table};
 use termtree::Tree;
 use crate::{common::*, constants::*, string_encode::decode_filename};
-
-#[derive(Tabled)]
-#[tabled(rename_all = "CamelCase")]
-pub struct TrashInfo {
-    #[tabled(rename = "Path")]
-    pub path: String,
-
-    #[tabled(rename = "Deletion Date")]
-    pub deletion_date: NaiveDateTime
-}
-
-impl TrashInfo {
-    fn from_file(path: PathBuf) -> Option<TrashInfo> {
-        let (trash_info_header, path_field, deletion_date_field) = (
-            String::from(TRASH_INFO_SECTION_HEADER).to_lowercase(),
-            String::from(TRASH_INFO_PATH_KEY).to_lowercase(),
-            String::from(TRASH_INFO_DELETION_DATE_KEY).to_lowercase()
-        );
-
-        match read_to_string(path) {
-            Ok(file_contents) => {
-                let mut ini = Ini::new();
-                match ini.read(file_contents) {
-                    Ok(map) => {
-                        match map.get(&trash_info_header) {
-                            Some(trash_info_section) => {
-                                let full_path = match trash_info_section.get(&path_field) {
-                                    Some(path_val) => path_val.clone(),
-                                    None => None
-                                };
-
-                                let deletion_date_str = match trash_info_section.get(&deletion_date_field) {
-                                    Some(deletion_date_val) => deletion_date_val.clone(),
-                                    None => None
-                                };
-
-                                if full_path.is_some() && deletion_date_str.is_some() {
-                                    let deletion_date = NaiveDateTime::parse_from_str(
-                                        deletion_date_str.unwrap().as_str(),
-                                        "%Y-%m-%dT%H:%M:%S"
-                                    );
-
-                                    match deletion_date {
-                                        Ok(date_result) => Some(TrashInfo { 
-                                            path: decode_filename(
-                                                Path::new(&full_path.unwrap())
-                                                    .file_name()
-                                                    .unwrap()
-                                                    .to_str()
-                                                    .unwrap()
-                                            ),
-                                            deletion_date: date_result
-                                        }) ,
-                                        Err(_) => None
-                                    }
-                                } else {
-                                    return None;
-                                }
-                            },
-                            None => None
-                        }
-                    },
-                    Err(_) => None
-                }
-            },
-            Err(_) => None
-        }
-    }
-}
 
 fn files_tree_label<P: AsRef<Path>>(p: P) -> String {
     let name = p.as_ref()
@@ -104,42 +38,6 @@ fn files_tree<P: AsRef<Path>>(p: P) -> Result<Tree<String>> {
     );
 
     Ok(result)
-}
-
-pub fn get_home_trash_contents() -> Result<Vec<TrashInfo>> {
-    let mut trash_contents = vec![];
-
-    match freedesktop_home_trash_info_dir() {
-        Some(info_path) => {
-            match read_dir(info_path) {
-                Ok(contents) => {
-                    for result in contents {
-                        match result {
-                            Ok(entry) => {
-                                let path = entry.path();
-                                if path.is_file() {
-                                    match TrashInfo::from_file(path) {
-                                        Some(trash_info) => trash_contents.push(trash_info),
-                                        None => continue
-                                    }
-                                }
-                            },
-                            Err(_) => continue
-                        }
-                    }
-
-                    Ok(trash_contents)
-                },
-                Err(error) => Err(error)
-            }
-        },
-        None => Err(
-            Error::new(
-                std::io::ErrorKind::Other, 
-                "Unable to determine the path for the home trash directory."
-            )
-        )
-    }
 }
 
 pub fn trash_list(recursive: bool) -> Result<()> {
